+++ python/xorbits/_mars/deploy/oscar/supervisor.py
@@ -74,13 +74,6 @@ class SupervisorCommandRunner(OscarCommandRunner):
         )
 
     async def start_services(self):
-        if (
-            "MARS_EXTERNAL_STORAGE" in os.environ
-            and "MARS_EXTERNAL_STORAGE" == "juicefs"
-        ):  # pragma: no cover
-            self.config["storage"]["backends"] = ["juicefs"]
-            self.config["storage"]["juicefs"]["root_dirs"] = ["/juicefs-data"]
-            self.config["storage"]["juicefs"]["is_k8s"] = True
         start_web = await start_supervisor(
             self.pool.external_address,
             self.args.supervisors,
