+++ python/xorbits/deploy/kubernetes/supervisor.py
@@ -12,12 +12,21 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import os
+
 from ..._mars.deploy.oscar.supervisor import SupervisorCommandRunner
 from .core import K8SServiceMixin
 
 
 class K8SSupervisorCommandRunner(K8SServiceMixin, SupervisorCommandRunner):
     async def start_services(self):
+        if (
+            "MARS_EXTERNAL_STORAGE" in os.environ
+            and "MARS_EXTERNAL_STORAGE" == "juicefs"
+        ):  # pragma: no cover
+            self.config["storage"]["backends"] = ["juicefs"]
+            self.config["storage"]["juicefs"]["root_dirs"] = ["/juicefs-data"]
+            self.config["storage"]["juicefs"]["is_k8s"] = True
         await super().start_services()
         await self.start_readiness_server()
         self.write_pid_file()
